Add erlyberly
erlyberly is not widespread (yet) . So, maybe should it not belong here. On the other hand it as some nice and advanced features that I like and other tools lack. Please, decide yourself, I will not be offended if you don't take it. (Hint: run the binary release for a try if you haven't used it yet.)